+++ sentential/lib/drivers/local_lambda.py
@@ -36,11 +36,11 @@ class LocalLambdaDriver(LambdaDriver):
         if credentials.SessionToken:
             credentials_env["AWS_SESSION_TOKEN"] = credentials.SessionToken
 
-        # TODO: remove if possible, now using docker desktop across platforms
-        # hosts = []
-        # # required to properly resolve `host.docker.internal` in Linux
-        # if platform == "linux":
-        #     hosts = [("host.docker.internal", "host-gateway")]
+        # required to properly resolve `host.docker.internal` when not using docker desktop.
+        # this isn't necessarily supported, but it does solve the problem.
+        hosts = []
+        if platform == "linux":
+            hosts = [("host.docker.internal", "host-gateway")]
 
         export_paths = ",".join([str(self.ontology.envs.path), str(self.ontology.secrets.path)])
 
@@ -52,6 +52,7 @@ class LocalLambdaDriver(LambdaDriver):
         clients.docker.run(
             image.id,
             name=LocalBridge.config.lambda_name,
+            add_hosts=hosts,
             hostname=LocalBridge.config.lambda_name,
             networks=[LocalBridge.config.bridge_name],
             detach=True,
